add check apply_project_list, participate_project_list
