+++ src/app/controller/mypage/getMypage.js
@@ -136,27 +136,5 @@ router.get('/:user_idx', async(req, res) => {
     }
 });
 
-//참여 및 지원한 프로젝트 모아보기
-router.get('/', async(req, res) => {
-    const ID = jwt.verify(req.headers.authorization);
-
-    if(ID != -1){
-        apply.aggregate([{'$group' : {'_id' : {'project_idx' : "$project_idx", 'join' : '$join'}}}], function(err, applies){
-            if(err) {
-                console.log(err);
-                return res.status(500).send({message: 'database failure'});
-            }
-            /*project.find({
-              '_id' : applies[0]
-            })*/
-            res.json(applies);
-        });
-    } else {
-        res.status(401).send({
-            message: "access denied"
-        });
-    }
-});
-
 
 module.exports = router;
